+++ tests/test_pandas.py
@@ -1,17 +1,20 @@
 """Test Pandas accessors."""
+import pytest
 import pandas as pd
 from redflag.pandas import SeriesAccessor, DataFrameAccessor
 
 
 c = pd.Series([1, 1, 1, 1, 1, 2, 2, 2, 3, 3])
-r = pd.Series([0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0])
+r = pd.Series([0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0, 3.0])
+
 
 def test_dummy_scores():
     c_scores = c.redflag.dummy_scores()
     r_scores = r.redflag.dummy_scores()
 
     assert c_scores['most_frequent']['roc_auc'] == 0.5
-    assert r_scores['mean']['mean_squared_error'] - 0.08249999999999999 < 1e-12
+    assert r_scores['mean']['mean_squared_error'] - 0.5710743801652893 < 1e-12
+
 
 def test_imbalance_metrics():
     minorities = c.redflag.minority_classes()
@@ -19,3 +22,19 @@ def test_imbalance_metrics():
 
     imb_degree = c.redflag.imbalance_degree()
     assert imb_degree - 1.25 < 1e-9
+
+
+def test_warnings():
+    with pytest.warns(UserWarning, match="The Series does not seem categorical."):
+        r.redflag.minority_classes()
+    with pytest.warns(UserWarning, match="The Series does not seem categorical."):
+        r.redflag.imbalance_degree()
+
+
+def test_series_categorical_report():
+    report_c = c.redflag.report()
+    assert 'Categorical' in report_c
+
+def test_series_continuous_report():
+    report_r = r.redflag.report()
+    assert 'Continuous' in report_r